# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Träwelling API",
        "description": "Träwelling user API description. This is an incomplete documentation with still many errors. The API is currently not yet stable. Endpoints are still being restructured. Both the URL and the request or body can be changed. Breaking changes will be announced on GitHub: https://github.com/Traewelling/traewelling/blob/develop/API_CHANGELOG.md",
        "contact": {
            "email": "support@traewelling.de"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0 - alpha"
    },
    "servers": [
        {
            "url": "https://traewelling.de/api/v1",
            "description": "Production Server"
        },
        {
            "url": "http://localhost:8000/api/v1",
            "description": "This instance"
        }
    ],
    "paths": {
...
        "/year-in-review": {
            "get": {
                "tags": [
                    "Statistics"
                ],
                "summary": "Returns the year in review for the given year and authenticated user",
                "description": "Please note: This endpoint is only available when the year in review feature is enabled in the backend configuration. There is no full documentation - this endpoint may change every year.",
                "operationId": "getYearInReview",
                "responses": {
                    "200": {
                        "description": "JSON object with the year in review data. The structure of the object may change every year. There is no full documentation at this point."
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Year in review is not active"
                    }
                },
                "security": [
                    {
                        "passport": [
                            "create-statuses"
                        ]
                    },
                    {
                        "token": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "OAuth2 authorizationCode Flow. Get your token from https://traewelling.de/settings/applications. Set the redirect URL to https://traewelling.de/api/oauth2-callback to use it here. See also https://laravel.com/docs/9.x/passport",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "authorizationCode": {
                        "authorizationUrl": "http://localhost:8000/oauth/authorize",
                        "tokenUrl": "http://localhost:8000/oauth/token",
                        "refreshUrl": "http://localhost:8000/auth/refresh",
                        "scopes": {
                            "read-statuses": "see all statuses",
                            "read-notifications": "see your notifications",
                            "read-statistics": "see your statistics",
                            "read-search": "search in Träwelling",
                            "write-statuses": "create, edit, delete statuses",
                            "write-likes": "create and remove likes",
                            "write-notifications": "mark notifications as read, clear notifications",
                            "write-exports": "request data exports",
                            "write-follows": "follow and unfollow users",
                            "write-followers": "accept follow requests and remove followers",
                            "write-blocks": "block and unblock users, mute and unmute users",
                            "write-event-suggestions": "suggest events in your name",
                            "write-support-tickets": "create support tickets in your name",
                            "read-settings": "see your settings, email, etc.",
                            "write-settings-profile": "edit your profile",
                            "read-settings-profile": "see your profile data, e.g. email",
                            "write-settings-mail": "edit your email",
                            "write-settings-profile-picture": "edit your profile picture",
                            "write-settings-privacy": "change your privacy settings",
                            "read-settings-followers": "show follow-requests and followers",
                            "write-settings-calendar": "create and delete new calendar-tokens",
                            "extra-write-password": "change your password",
                            "extra-terminate-sessions": "log you out of other sessions and apps",
                            "extra-delete": "delete your account"
                        }
                    }
                }
            },
            "token": {
...
    },
...
}